# backend/internal/transport/http/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Racion API",
    "version": "1.0.0",
    "summary": "Weekly meal plans with real store prices and one shopping list.",
    "description": "Read-only catalogue plus plan building. No key, no account: send a request, get JSON.\n\nThe planner is deterministic — the same answers give the same week, and every dish carries the reason it was chosen. Prices are estimates from national statistics (Rosstat in Russia, Eurostat price levels with HICP in Europe, BLS in the US) multiplied by a store index; they are never live shelf prices and must be presented as estimates.\n\nRecipes and photos are ours and licensed for personal, non-commercial use. Building a product on top of this API is fine; republishing the recipe corpus is not.\n\nRate limit per IP address: 600 requests a minute to /api/ overall, of which 60 a minute may build plans (bursts of 100 and 20 are allowed). Over the limit the answer is 429 with a Retry-After header. CORS is open, so the API can be called straight from a browser. Language comes from ?lang=; prices from ?country= and the chosen store.",
    "contact": {
      "name": "Racion",
      "email": "info@racion.app",
      "url": "https://racion.app"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    }
  },
  "servers": [
    {
      "url": "https://racion.app",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/api/meta": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "operationId": "getMeta",
        "summary": "Countries, stores, filters and allergens",
        "description": "Everything needed to fill in plan parameters: the list of supported countries with currency and decimals, the stores of one country with their price index, catalogue filters, allergens and exclusion presets.",
        "parameters": [
          {
            "$ref": "#/components/parameters/country"
          },
          {
            "$ref": "#/components/parameters/lang"
          }
        ],
        "responses": {
          "200": {
            "description": "Reference data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "operationId": "listRecipes",
        "summary": "Recipe catalogue with search and paging",
        "description": "Returns short cards: title, meal, time, calories and the cost of one portion in the currency of ?country=. Use ?q= for search, ?slot= to keep one meal only.",
        "parameters": [
          {
            "$ref": "#/components/parameters/country"
          },
          {
            "$ref": "#/components/parameters/lang"
          },
          {
            "name": "q",
            "in": "query",
            "description": "Free-text search over titles and ingredients",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slot",
            "in": "query",
            "description": "Keep one meal only",
            "schema": {
              "type": "string",
              "enum": [
                "breakfast",
                "lunch",
                "dinner",
                "snack"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1–200",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of recipe cards",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipePage"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/{id}": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "operationId": "getRecipe",
        "summary": "One recipe with ingredients, steps and nutrition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "syrniki"
          },
          {
            "$ref": "#/components/parameters/country"
          },
          {
            "$ref": "#/components/parameters/lang"
          }
        ],
        "responses": {
          "200": {
            "description": "The recipe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/collections": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "operationId": "listCollections",
        "summary": "Curated collections with their recipe ids",
        "description": "Holiday tables, quick dinners, pizza doughs, childhood food and the rest. A collection id can be passed to POST /api/plans as `collection` so the planner takes its dishes first.",
        "parameters": [
          {
            "$ref": "#/components/parameters/lang"
          }
        ],
        "responses": {
          "200": {
            "description": "Collections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Collection"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/occasions": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "operationId": "listOccasions",
        "summary": "Holiday tables that can be built for a number of guests",
        "parameters": [
          {
            "$ref": "#/components/parameters/lang"
          }
        ],
        "responses": {
          "200": {
            "description": "Occasions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Occasion"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ingredients": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "operationId": "listIngredients",
        "summary": "Ingredient reference: units, pack sizes, nutrition per 100 g",
        "parameters": [
          {
            "$ref": "#/components/parameters/lang"
          }
        ],
        "responses": {
          "200": {
            "description": "Ingredients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ingredient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans": {
      "post": {
        "tags": [
          "plans"
        ],
        "operationId": "createPlan",
        "summary": "Build a week of meals with a shopping list",
        "description": "The main call. Give who eats, where they shop and what they cannot have; get seven days of dishes, per-day nutrition, and one shopping list rounded to whole packs with an estimated total.\n\nThe result is stored and readable at GET /api/plans/{id}; the same id opens the human page at /plan/{id}, which is what you show a person.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Params"
              },
              "examples": {
                "family": {
                  "summary": "Two adults and a child, Пятёрочка, no fish",
                  "value": {
                    "country": "RU",
                    "store": "pyaterochka",
                    "region": "",
                    "adults": 2,
                    "kids": [
                      {
                        "age": 7
                      }
                    ],
                    "slots": [
                      "breakfast",
                      "lunch",
                      "dinner"
                    ],
                    "equipment": [
                      "stove",
                      "oven"
                    ],
                    "excludeTags": [
                      "fish"
                    ],
                    "allergens": [],
                    "budgetMode": "perPersonDay",
                    "budgetValue": 400,
                    "lang": "ru"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadInput"
          },
          "429": {
            "$ref": "#/components/responses/TooMany"
          }
        }
      }
    },
    "/api/plans/{id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "operationId": "getPlan",
        "summary": "Read a plan that was built earlier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/lang"
          }
        ],
        "responses": {
          "200": {
            "description": "The plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/occasions/{id}": {
      "post": {
        "tags": [
          "plans"
        ],
        "operationId": "createOccasionPlan",
        "summary": "Build a holiday table for a number of guests",
        "description": "One day instead of a week: the planner fills the courses of the occasion (starters, mains, salads, dessert) for the given number of guests.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "newyear"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "guests",
                  "params"
                ],
                "properties": {
                  "guests": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 40
                  },
                  "params": {
                    "$ref": "#/components/schemas/Params"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The table",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadInput"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "country": {
        "name": "country",
        "in": "query",
        "description": "Two-letter country code: prices and currency come from it. RU by default.",
        "schema": {
          "type": "string",
          "example": "RU"
        }
      },
      "lang": {
        "name": "lang",
        "in": "query",
        "description": "Interface and recipe language, one of 15 codes. ru by default.",
        "schema": {
          "type": "string",
          "enum": [
            "ru",
            "en",
            "de",
            "fr",
            "es",
            "it",
            "pt",
            "pl",
            "uk",
            "tr",
            "nl",
            "cs",
            "kk",
            "zh",
            "ja"
          ]
        }
      }
    },
...
  }
}